destroyアクション条件分岐追加

+++ app/controllers/items_controller.rb
@@ -1,7 +1,7 @@
 class ItemsController < ApplicationController
   before_action :authenticate_user!, except: [:index, :show]
-  before_action :set_item, only: [:edit, :show, :update]
-  before_action :move_to_index2, except: [:create, :new, :index, :show, :destroy]
+  before_action :set_item, only: [:edit, :show, :update, :destroy]
+  before_action :move_to_index2, except: [:create, :new, :index, :show,]
   
   def index
     @items = Item.order("created_at DESC")
@@ -18,8 +18,12 @@ class ItemsController < ApplicationController
   end
 
   def destroy
-    if @item.destroy
-      redirect_to root_path
+    if current_user.id == @item.user_id
+      if @item.destroy
+        redirect_to root_path
+      else
+        render :show
+      end
     end
   end
